List dependencies for services - add integration test to verify
