+++ moto/stepfunctions/models.py
@@ -4,9 +4,8 @@ from datetime import datetime
 
 from boto3 import Session
 
-from moto.core import BaseBackend
+from moto.core import ACCOUNT_ID, BaseBackend
 from moto.core.utils import iso_8601_datetime_without_milliseconds
-from moto.sts.models import ACCOUNT_ID
 from uuid import uuid4
 from .exceptions import (
     ExecutionAlreadyExists,
