+++ requirements-dev.txt
@@ -1,17 +1,15 @@
 -r requirements.txt
-nose
+-r requirements-tests.txt
+
 black==19.10b0; python_version >= '3.6'
 regex==2019.11.1; python_version >= '3.6'   # Needed for black
-sure==1.4.11
 coverage==4.5.4
 flake8==3.7.8
-freezegun
 flask
 boto>=2.45.0
 boto3>=1.4.4
 botocore>=1.15.13
 six>=1.9
-parameterized>=0.7.0
 prompt-toolkit==2.0.10 # 3.x is not available with python2
 click==6.7
 inflection==0.3.1
