+++ requirements-tests.txt
@@ -1,0 +1,4 @@
+nose
+sure==1.4.11
+freezegun
+parameterized>=0.7.0
