+++ tests/test_autoscaling/test_autoscaling_cloudformation.py
